Renamed the default index file to be index1.js, and added custom code to a new index.js file for the landing page.
